Configure esy monorepo
The idea is that `esy.json` links to all packages in the monorepo.

For each `<pkg>` it has:

    "dependencies": {
      "<pkg>": "*",
      ...
    },
    "resolutions": {
      "<pkg>": "link:./<pkg>.json",
      ...
    }

Also we add a little helper `esydune` which configures `dune` given the
esy environment for the package. Thus each package has

    "esy": {
      "build": "esydune build",
      "build": "esydune install"
    }

as its `"esy"` configuration.

We also remove `"devDependencies"` configuration from all packages but
the root one (`esy.json`). As we are going to develop packages as a
whole we want to configure `"devDependencies"` only ath the root package
of the monorepo and then make them available to each package's build
environment.

Thanks for composable dune CLI we even can invoke aarbitrary `dune` commands
that way and they will be scoped for the current package rather than for
the whole dune project (which don't want to use because we are using
esy-way of stitching packages together):

    % esydune runtest
    esydune: dune runtest --root . --only-packages "$cur__name"