Add esymono command
This commit adds `esymono` command which wraps `esy` to configure monorepo
workflow.

The main reason we need this is because we want `"devDependencies"` configured
at the root package's `esy.json` be available to each package's build
environment. Therefore we represent common esy commands (`esy build`, ...) via
esy low level commands invocation which sets a dependency specification
(DEPSPEC).

For example this command will build all dependencies in the monorepo
providing root's devDeps to each package build environment:

    % esy build-dependencies \
        --link-depspec 'dependencies(self)+devDependencies(root)' \
        --all \
        root

Or via `esymono`:

    % esymono build

We can also build all monorepo in "release" mode (no devDependencies provided
and therefore we simulate the env in which packages will be built when
released):

    % esymono build --release

We can build only a specific package by referring to it by its name:

    % esymono build @reason-native/rely

Or we can refer by the manifest:

    % esymono build ./rely.json

Which maybe more convenient as shell autocompletes paths by default.

To execute an arbitrary `COMMAND` in a dev environment w/o waiting for packages
in the monorepo to be built:

    % esymono vim .

To execute an arbitrary `COMMAND` in a dev environment but before ensuring that
packages are built:

    % esymono x TestPastel.exe

So this can be used to run tests.

It's also possible to narrow the environment to the specific package:

    % esymono x-in ./rely.json echo '#{self.name}'