Extract reporter API, change reporter output to be more jest-like (#63)
Summary:
I decoupled the reporting behavior from the core test itself by exposing a set of life cycle methods that are analogous to those Jest exposes. I ended up changing the terminal output to be more Jest-like as well because the previous behavior was not fully compatible with those methods, and the goal is to avoid bikeshedding by deferring to Jest when it comes to output.

The new output can be seen below.

![image](https://user-images.githubusercontent.com/5252755/50254055-6f186b00-03a1-11e9-9187-12e3c2fe2ab9.png)

![image](https://user-images.githubusercontent.com/5252755/50254223-23b28c80-03a2-11e9-8f5f-f924e372576b.png)

![image](https://user-images.githubusercontent.com/5252755/50254086-90795700-03a1-11e9-9447-739d26c5e076.png)
Pull Request resolved: https://github.com/facebookexperimental/reason-native/pull/63

Reviewed By: jordwalke

Differential Revision: D13530295

Pulled By: jordwalke

fbshipit-source-id: 278e4023a7303bac52cc0e80c40db0782acd02ac