Benderson/collection matchers (#70)
Summary:
Adds expect.list and expect.array matchers
![image 8](https://user-images.githubusercontent.com/5252755/50527266-25640a80-0a9c-11e9-8855-8b144505ebde.png)
![image 9](https://user-images.githubusercontent.com/5252755/50527271-2c8b1880-0a9c-11e9-9bc3-6ca1a999110c.png)
![image 10](https://user-images.githubusercontent.com/5252755/50527280-33b22680-0a9c-11e9-8b28-f103286b717b.png)
![image 11](https://user-images.githubusercontent.com/5252755/50527283-357bea00-0a9c-11e9-9fa9-fabbcbeb3d99.png)

Targeting master because I can't get a clean diff to custom matchers because of merge conflicts/rebases. You can look at individual commits on 12/27 - 12/28 to review.

Main point of contention here is likely to be the ~equals argument (which defaults to == which has problems with floats/functions) to toEqual and toContain. Internally we have been using a variant that handles Float behavior and has a Custom constructor for custom equality. My opinion is that the variant added verbosity that was generally not needed and problems should make themselves known pretty quickly and be resolved fairly intuitively. It would be cool if we could keep the current API but be intelligent about providing a default for Floats and functions, but I don't really know how to accomplish that.
Pull Request resolved: https://github.com/facebookexperimental/reason-native/pull/70

Differential Revision: D13846653

fbshipit-source-id: 9569e22b36eb72b4f701f8553ed8354377d32c93