Make reporters API public (#98)
Summary:
See RunConfig.re and Rely.rei for the main changes. Basically custom reporters can be specified via a Custom constructor to RunConfig (and the default reporter is just called Default, I could see calling it terminal as well).

I played with a few solutions for exposing the test result types, I ended up aliasing the types we needed directly on the Reporter module and not exposing TestResult directly. I played with separating the public types from the intermediate types and conversion code, but wasn't super happy with how this ended up. I could see separating TestResult into three intermediate modules and requiring a limited signature in Reporter as well, but wanted to discuss here before commiting to that.

I want to say that this is a minor version, but we are no longer exposing the TestResult module on Rely.rei, which is a breaking change. That said it was only ever usable for reporters which were explicitly marked as internal soooo /shrug.
Pull Request resolved: https://github.com/facebookexperimental/reason-native/pull/98

Differential Revision: D13991514

fbshipit-source-id: d17531cf7197a8d0f400377796390fea75e42933