Benderson/pastel infer mode (#102)
Summary:
Addresses https://github.com/facebookexperimental/reason-native/issues/100

- Checks Unix.isatty and the FORCE_COLOR environment variable, and sets the default mode of Pastel accordingly
  - FORCE_COLOR takes precedence
  - if FORCE_COLOR is undefined, then if non tty we disable by default
- Update Rely Terminal reporter to not use Ansi escape characters to erase lines if Pastel.mode != terminal (only prints after a suite has run instead of at beginning in that case)
- Pass FORCE_COLOR true to refmterr tests to not break snapshots
  - There may be a better way to do this than what I did
- Update rely snapshots (the change to terminal reporter behavior broke some old matcher tests)

Ultimately I chose to stop short of publicly exposing supportsColor and inferMode as I think we will want to flush it out a fair bit more (compare to behavior of chalk/supports-color) in terms of things we check as well as modifying our behavior in 16 bit/256 bit/true color support terminals and really figuring out what we want our API to be.

I am not sure if this is a breaking change for Pastel or not and would like to talk that over, once merged I will cut a new version of Pastel (including the powershell support work), and will cut a Rely release as well once Mock matchers are merged to avoid a breaking change with the Mock API that matchers introduced (Rely.Mock hasn't been released yet).

![image](https://user-images.githubusercontent.com/5252755/52888905-5d99d800-3132-11e9-9594-8e3be64e8a60.png)
Pull Request resolved: https://github.com/facebookexperimental/reason-native/pull/102

Differential Revision: D14127085

fbshipit-source-id: 538dbb1a78b3464a09c4d7c84c9d7bb41844408c