Benderson/mock matchers (#101)
Summary:
Implement matchers for Rely.mock

The actual MockMatchers.re and MockMatchers_test.re need to be manually expand because they are too large to be rendered by default.

MatcherSnapshotTestRunner.re is also worth checking out as a better way for writing snapshot tests for matcher output leveraging the now exposed Reporter API (and converting stuff over could be bootcamped internally/added as an issue here)

I had to do some finagling to get types to flow through correctly, I don't love that the mock type itself isn't in Mock.Sig, but that clearly needs to be done now to prevent having to do other breaking changes (such as moving Describe and Test modules into the Rely.Make functor). We already know that with the next major version we will be doing some consolidation into the Rely.Make functor, but I am less sure as to what belongs there and what does not, with the main concern being the eventual composability of test projects/having a top level executable for dynamically loading and executing test libraries.
Pull Request resolved: https://github.com/facebookexperimental/reason-native/pull/101

Differential Revision: D14127093

fbshipit-source-id: 92e3edc54449fbcf032879aa0d7677522b048e1b