Enable color detection on Mingw/Cygwin/Mintty shells.
Summary:Unix.is_atty returns false for mintty - kind of a longstanding
bug. This fixes that. It's actually not easy to detect whether or not
you are in a cygwin/mintty shell. I used an open source project as a
starting point.

This also would have fixed our previous bug of failing in the c stub for
setting console mode. But this way we have double coverage.

Eventually this should all go into a standalone library for terminal
capabilities detection.

Test Plan:Will let CI run, and also tested on powershell as well as
mintty shell.

Reviewers:kad

CC: