Enable color detection on Mingw/Cygwin/Mintty shells. (#113)
Summary:
Unix.is_atty returns false for mintty - kind of a longstanding
bug. This fixes that. It's actually not easy to detect whether or not
you are in a cygwin/mintty shell. I used an open source project as a
starting point.

This also would have fixed our previous bug of failing in the c stub for
setting console mode. But this way we have double coverage.

Eventually this should all go into a standalone library for terminal
capabilities detection.
Pull Request resolved: https://github.com/facebookexperimental/reason-native/pull/113

Differential Revision: D14215898

fbshipit-source-id: de40624da1b1fe18e3a1b27447ab3b553d0bdffb