Add location to Passed and Skipped TestStatus Constructors for CI purposes
Summary:
This is technically a breaking change, but I think a necessary one/a general improvement for anyone using custom reporters.

I don't plan on cutting a release of Rely with this until we also get in some of the top level module restructuring (e.g. changes to accomodate testSkip/testOnly, moving more things inside RHS of  Rely.Make functor)

Reviewed By: kyldvs

Differential Revision: D14423114

fbshipit-source-id: 7ae65b81b1eb0f280f63feae3c9c9a11416e2033