Separate out rely internal to allow testing without adding to the public API (#130)
Summary:
Now that we are batching together a set of backwards breaking changes for Rely 2.0.0, I think it is a goods time to implement this.

Basically I have been frustrated in unit testing some components due to needing to add them to the public API (most notably collection matchers). I can see this continuing in the future, and to get around this I propose moving essentially all of Rely's logic into a separate library (rely.internal) that rely.lib and our tests project depend on. I have implemented that in this diff and also removed collection matchers internals from the public API as a proof of concept.

Consumers will not need to change their dependencies, and unless they depend on collection matcher internals for some reason, should not be affected by this change (though it is of course technically breaking, which is why it is being batched with the other breaking changes we wish to make)
Pull Request resolved: https://github.com/facebookexperimental/reason-native/pull/130

Differential Revision: D14581164

fbshipit-source-id: 97c91b9db9e68a5a0e638ed0debb9553ad663c6b