Test and describe only (#151)
Summary:
Addresses https://github.com/facebookexperimental/reason-native/issues/81

This PR implements testOnly and describeOnly as well as ciMode (which can be specified manually via run config or by passing the --ci flag).

The behavior of testOnly and describeOnly diverges (intentionally) from Jest and works in the following way:

Anything marked as only or with an ancestor (inside a describe or nested describe) marked as only will run as long as it does not have an ancestor marked as skipped (skip takes precedence over only)

As long as any test/describe being run is marked as only, ONLY tests/describes that are marked as only or with ancestors marked as only will run.

If anything is marked as only in CI mode, an exception is thrown that looks like the following

![image](https://user-images.githubusercontent.com/5252755/56393001-e3e0a200-61e7-11e9-9b45-a7fd8c65a90f.png)
Pull Request resolved: https://github.com/facebookexperimental/reason-native/pull/151

Differential Revision: D15007041

Pulled By: bandersongit

fbshipit-source-id: c7fa2e96e5c17f1342ddbe78c6163d33145c0209