Remove remaining usage of Str in Rely, Pastel, and Common (#165)
Summary:
I've been working on getting Rely to work with JSOO (spoiler alert, I succeeded with these changes) and Str was one of the blocking issues (and we wanted to remove it anyway). This PR removes the reamining usages of Str in Rely, Pastel, and Common.

I also added //Provides: comments in the Pastel js stubs so that they correctly get picked up by JSOO.

There was a small change in behavior due to the behavior of Str.split_delim/Str.full_split vs. Re.split_full. I implemented a split_delim in the case we actually relied on it, and in the other cases the change was non functional and I think was better than before (you can see some snapshots where we basically had the empty string inside of a default Pastel get removed)
Pull Request resolved: https://github.com/facebookexperimental/reason-native/pull/165

Reviewed By: jordwalke

Differential Revision: D15386118

Pulled By: bandersongit

fbshipit-source-id: 4597ba37baaebcec7611504e978790663fcdfbc0