Migrate refmterr tests to Rely (#176)
Summary:
Hey,

I spend some time migrating the `refmterr` tests to Rely in order to learn more about Rely and reason-native. It's been just a few weeks since I started playing with ReasonML, so feel free to give me any tip. I started with [this issue#25](https://github.com/facebookexperimental/reason-native/issues/25) since I wanted to help reason-native move forward.

This PR includes the changes related to removing the test-runtime that were testing the `refmterr` cli. I tried to not change how the tests are made and focus on have them running with Rely and in next iterations we could improve those.

I needed to change the mode in of `Pastel` to `Terminal`. Without this, the encoding wasn't working properly. I could enable that only in testing but doesn't seem a proper test if there's a difference on the output of the console in testing. I'm not 100% sure of this change, so, if there's something wrong with it I might need help.

For instance, I would like to refactor those tests into a shape similar to how https://github.com/satya164/babel-test manages the fixtures. The thing that I like most is having a little code on the test-side and being the folder structure that gives all the information to run the test-cases.

We might not want that, but would be necessary at some point to separate `specialTests` and `textTests` from all the rest. It would have more or less that structure: https://gist.github.com/davesnx/dac795d4d2c518995fefbff822c3e4af

The other improvement that we could do afterwards is to move the try and the rm cmi's into a `beforeEach` so, I'm waiting for https://github.com/facebookexperimental/reason-native/issues/171.
Pull Request resolved: https://github.com/facebookexperimental/reason-native/pull/176

Reviewed By: jordwalke

Differential Revision: D15742773

Pulled By: bandersongit

fbshipit-source-id: bdb0dd41cb6489b9f29478f6aba3b906c47a002b