add CLI only option to only print details for failed suites (#179)
Summary:
For now I just added this as a CLI only option as it seemed the most lightweight solution that addresses our immediate internal needs.

I don't really want to touch RunConfig for now and there are other ways of dealing with the "lots of input problem" (filters, some heuristic about when to switch to this behavior, looking at source control/dependency graphs) that we may want to pursue in favor of this in the future.

Tested by adding/removing the flag from our TestDev command (left it added for better DevX)
Pull Request resolved: https://github.com/facebookexperimental/reason-native/pull/179

Reviewed By: kyldvs

Differential Revision: D15909783

Pulled By: bandersongit

fbshipit-source-id: 1e4a730e0b695b8a312447da584eebfc441c8d86