Add ability to partition a decorated string based on the unformatted text (#183)
Summary:
kyldvs this is to address the issue you had when trying to support colored strings in your string table library. Basically you can now use Pastel.partition to split a string into two self contained parts (e.g. if you split some string `s` into `s1` and `s2`, you can insert some stuff between `s1` and `s2` and  the decorators from s1/s2 will not be applied, but `s1` and `s2` are decorated as they were in `s`

I added an unformattedText function as well to aid in testing this and felt that it was reasonable to leave in the public API. I also did a fair bit of refactoring to consolidate some logic.
Pull Request resolved: https://github.com/facebookexperimental/reason-native/pull/183

Differential Revision: D16187279

Pulled By: bandersongit

fbshipit-source-id: 11c09ac6e6aee3bb5f0c83ecf01ef1a19ed895cb