RFC: Rename Path library to Pth
Summary:
The core compiler libs have a library named Path, and it can conflict with this one since we picked such a common name that was not 'namespaced' inside of a larger library name. `Pth` doesn't seem so bad as an alternative and kind of matches `Fs`. It's one character shorter which is nice, but it's slightly less clear.
Let me know if you think this is the wrong call.

Reviewed By: kyldvs

Differential Revision: D17281706

fbshipit-source-id: f992f01a00393cdc88724954cc98d0ba04b9c6ce