Remove pth.install from esy.json
Summary: In a previous commit, the `Path` file was renamed. It was initially renamed to `pth` before a discussion that changed it to `fp`. The `pth.install` reference was missed in `esy.json` during the rename, which is breaking testing. This diff updates `esy.json` to use `fp.install` instead.

Reviewed By: kyldvs

Differential Revision: D17743127

fbshipit-source-id: e081f49d2aa3315797d1233cfe31962b24280c51