Add Table formatting to Console Module (#203)
Summary:
This PR adds table printing to the Console Module. It allows for creating a table with column widths, border type, and border style in Reason JSX. It will also allow for Pastel to be used in a table cell (although doesn't support nested Pastel tags).

The table components are in the `Console.ConsoleTable` module. See the test cases for examples of their use.

With respect to the formatting, I went with the simple solution of creating a new set of dividers for each outer and inner border type. If we add many more styles in the future, it would be worthwhile to move this logic into the table component rather than the border style module.
Pull Request resolved: https://github.com/facebookexperimental/reason-native/pull/203

Test Plan:
Imported from GitHub, without a `Test Plan:` line.

Ran included Rely tests.

Differential Revision: D17886237

Pulled By: evanyeung

fbshipit-source-id: ff97c6abe3b93ccc7f8a9f20181eaf4ab89455c5