fix opamify version bug for depending on exact versions (#244)
Summary:
Tested manually with common version patterns and rerunning esy-prepublish.

jordwalke did you have any test cases for this/I imagine this issue will be present in a bunch of other repos. Can you tell me where else this should be fixed/do you think it would be a good idea to package all this logic up so these fixes are easier to address in the future?
Pull Request resolved: https://github.com/facebookexperimental/reason-native/pull/244

Differential Revision: D20142648

Pulled By: bandersongit

fbshipit-source-id: 21fe592e1d71144363fe1e3689fae752e2fd8bed