Generate opam file for cli (#248)
Summary:
This adds an opam file for `cli`.

I'd like to use Rely with Opam, and even though it's not released on Opam yet, I can use `opam pin` as a workaround. Problem is, I also need to pin all of Rely's dependencies in the reason-native repository. As `cli` is one of them, I would need to be able to pin it, thus the need for an opam file.

I haven't changed the json file much, it's mostly indentation, but I have added a `description` field as it is required by `esy-prepublish.js`.
Pull Request resolved: https://github.com/facebookexperimental/reason-native/pull/248

Differential Revision: D20355326

Pulled By: bandersongit

fbshipit-source-id: 957b31621d15e08627c1f92778a99f0d85839653